+++ RAK11200BatteryCharge/Properties/AssemblyInfo.cs
