+++ RAK11200BatteryCharge/packages.config
