+++ RAKWisBlock.sln
@@ -23,7 +23,7 @@ Project("{11A8DD76-328B-46DF-9F39-F559912D0360}") = "AzureIoHubRAK1901HttpSasTok
 EndProject
 Project("{11A8DD76-328B-46DF-9F39-F559912D0360}") = "AzureIoHubRAK1901HttpSasKey", "AzureIoHubRAK1901HttpSasKey\AzureIoHubRAK1901HttpSasKey.nfproj", "{FB2079FF-F5F3-4FFA-B874-5BDB1A41E7E4}"
 EndProject
-Project("{11A8DD76-328B-46DF-9F39-F559912D0360}") = "RAK11200BatteryVoltage", "RAK11200BatteryVoltage\RAK11200BatteryVoltage.nfproj", "{C9EB6D89-11B8-4664-9FE5-9E1980D03C5A}"
+Project("{11A8DD76-328B-46DF-9F39-F559912D0360}") = "RAK11200BatteryCharge", "RAK11200BatteryCharge\RAK11200BatteryCharge.nfproj", "{C9EB6D89-11B8-4664-9FE5-9E1980D03C5A}"
 EndProject
 Project("{11A8DD76-328B-46DF-9F39-F559912D0360}") = "AzureIoTHubRAK11200PowerConservation", "AzureIoTHubRAK11200PowerConservation\AzureIoTHubRAK11200PowerConservation.nfproj", "{BD33B036-2759-4B02-8B4A-02DFACA2D00C}"
 EndProject
